+++ djangopypi2/apps/pypi_frontend/management/commands/ppadd.py
@@ -19,7 +19,7 @@ from urlparse import urlsplit
 from setuptools.package_index import PackageIndex
 from django.contrib.auth.models import User
 from django.utils.datastructures import MultiValueDict
-from ....pypi_config.models import Classifier
+from ....pypi_metadata.models import Classifier
 from ...models import Package
 from ...models import Release