+++ Changelog.rst
@@ -1,5 +1,8 @@
 History
 =======
+0.5.5 (2012-10-07)
+------------------
+* New pypi_metadata app, holds only package metadata
 
 0.5.4 (2012-10-05)
 ------------------